Allom Uncert: Aggregate to plot level first
We can preserve the allometric uncertainty at the plot level, so to
make sure we’re doing things right, we should not go from the tree to
the site level.  Instead we need to go from trees to the plot and then
plot to site, while preserving the variation from the allometric
uncertainty (which isn’t a problem).

Note: This means that the increment uncertainty is an oddball and the
only one where we ignore plot-level variation.  This is because the
increment uncertainty comes from the aggregating trees to plot step and
there is no way to do the proper tree -> plot -> site progression while
preserving the increment uncertainty (or at least I can’t figure it
out).  So for the purposes of this paper, we’re treating each site as 1
big plot.