Initialize the Branch for all Sites
All dated sites have been added to the RWL/RWL_alltrees.rwl and the
raw_input_files/core.data file.  Made this its own branch so as not to
mess with the Uncertainty paper.  I can then work from here to make any
changes we need to the scripts to look at Biomass for the whole site.