Increment deviations from mean
set up scripts to begin adding the different areas of uncertainty
together in quadrature.  This includes taking random samples of
deviations from the different area of uncertainty and adding them to
the baseline mean.  Coded to include the allom.array right now, but
will add other areas in the near future.