Stacked Uncertainty Recalc
Recalculated how uncertainties are stacked for the increment figure.
Now have graphs broken down by area.  Similar to that of the cumulative
figure that shows the relative contribution of each area of uncertainty
to the overall uncertainty at each time step.  @crollinson, I
referenced the increment deviations off of the mean(uncertainty area
increment).  This is similar to what we did for the cumulative figure,
but I would like to double check it with you.