Corrected order of Operations
Corrected the order of operations issues we were having in #14.
Calculating the 95% CI off of the deviations of each area of
uncertainty from the mean of that respective area.  THEN, adding up the
upper CI and lower CI from each area of uncertainty to get the total
Upper CI and total lower CI.

#15—Added these newly added biomass inc. CI’s to the climate
correlations.  Using the deviations from the mean that have been added
together, and am not referencing off of the overall mean as we did
previously.  This also shows a more informative story in #16 when we
plot the correlations of each indiv. area of uncertainty with climate.